// src/app/models/material.ts
export class Material {
    mid: string;
    projectid: string;
    name: string;
    
    brandid: string;  //need a CRUD screen for brand
    brand: string;

    price: number;
    price_escalation: number;
    prime_rate: number;
    accessory_rate: number;
    
    discount: number;
    
    
    
    
    created_date: string;
    created_by: string;
    updated_date: string;
    updated_by: string;

    // constructor(name, price, brandid, brand, projectid, discount, prime_rate, price_escalation, accessory_rate, created_by, created_date, updated_by, updated_date) { 
    //     this.name = name;
    //     this.price = price;
    //     this.price_escalation = price_escalation;
    //     this.brandid = brandid;
    //     this.brand = brand;
    //     this.projectid = projectid;
    //     this.discount = discount;
    //     this.prime_rate = prime_rate;
    //     this.accessory_rate = accessory_rate;
    //     this.created_date = created_date;
    //     this.created_by = created_by;
    //     this.updated_date = updated_date;
    //     this.updated_by = updated_by;
    // }
}